refactor(buttondown): share snapshot validation
